- refactored actors cards
